+++ src/war_state_machine_chopping.c
@@ -14,7 +14,7 @@ void enterChoppingState(WarContext* context, WarEntity* entity, WarState* state)
     vec2 position = getUnitCenterPosition(entity, true);
     vec2 treePosition = state->chop.position;
 
-    setStaticEntity(map->finder, position.x, position.y, unitSize.x, unitSize.y, entity->id);
+    setStaticEntity(map->finder, (s32)position.x, (s32)position.y, (s32)unitSize.x, (s32)unitSize.y, entity->id);
     setUnitDirectionFromDiff(entity, treePosition.x - position.x, treePosition.y - position.y);
     setAction(context, entity, WAR_ACTION_TYPE_HARVEST, true, 1.0f);
 }
@@ -41,7 +41,7 @@ void updateChoppingState(WarContext* context, WarEntity* entity, WarState* state
     }
 
     vec2 treePosition = state->chop.position;
-    WarTree* tree = getTreeAtPosition(forest, treePosition.x, treePosition.y);
+    WarTree* tree = getTreeAtPosition(forest, (s32)treePosition.x, (s32)treePosition.y);
 
     if (!tree || tree->amount == 0)
     {
